add qsvt notebook
