+++ algorithms/hamiltonian_simulation/hamiltonian_simulation_with_block_encoding/hamiltonian_simulation_qsvt.metadata.json
@@ -1,0 +1,9 @@
+{
+  "title": "Hamiltonian Simulation with QSVT",
+  "subtitle": "Hamiltonian Simulation with Quantum Singular Value Transformation",
+  "description": "Implementation of time-independent Hamiltonian simulation via the Quantum Singular Value Transformation (QSVT) method. The algorithm decomposes the time evolution as cos(Ht) - i*sin(Ht), approximates each function separately as Chebyshev polynomials via the Jacobi\u2013Anger expansion, and combines them via an optimized LCU using the qsvt_lcu function. QSVT achieves optimal query complexity without controlled block-encoding calls, using two auxiliary qubits.",
+  "friendly_name": "Hamiltonian Simulation with QSVT",
+  "problem_domain_tags": [],
+  "qmod_type": [],
+  "vertical_tags": []
+}
